+++ src/components/TextRangeSelector/Renderer/BackgroundLayer.tsx
