+++ src/components/TextRangeSelector/Renderer/Cursor.tsx
