+++ src/components/TextRangeSelector/Renderer/index.tsx
