+++ src/components/TextRangeSelector/index.tsx
@@ -1,5 +1,5 @@
 import { useContext, useEffect } from "react";
-import TextRender from "./TextRender";
+import TextRender from "./Renderer";
 import { NewTRSContext, NewTRSProvider } from "./context/NewTRSContext";
 
 function Render({ text }: { text: string }) {
